+++ cpukit/rtems/include/rtems/rtems/msgmp.h
@@ -65,7 +65,7 @@ typedef struct {
   rtems_option                       option_set;
   Objects_Id                         proxy_id;
   uint32_t                           count;
-  uint32_t                           size;
+  size_t                             size;
   uint32_t                           pad0;
   CORE_message_queue_Buffer          Buffer;
 }   Message_queue_MP_Packet;
@@ -99,7 +99,7 @@ rtems_status_code _Message_queue_MP_Send_request_packet (
   Message_queue_MP_Remote_operations  operation,
   Objects_Id                          message_queue_id,
   void                               *buffer,
-  uint32_t                           *size_p,
+  size_t                             *size_p,
   rtems_option                        option_set,
   Watchdog_Interval                   timeout
 );
