+++ cpukit/rtems/src/msgqbroadcast.c
@@ -56,7 +56,7 @@
 rtems_status_code rtems_message_queue_broadcast(
   Objects_Id            id,
   void                 *buffer,
-  uint32_t              size,
+  size_t                size,
   uint32_t             *count
 )
 {
